Move bst package into logic
