+++ src/main/java/net/sf/jabref/logic/bst/VM.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.bst;
+package net.sf.jabref.logic.bst;
 
 import java.io.File;
 import java.io.IOException;
