+++ src/test/java/net/sf/jabref/logic/bst/BibtexCaseChangersTest.java
@@ -1,6 +1,6 @@
-package net.sf.jabref.bst;
+package net.sf.jabref.logic.bst;
 
-import net.sf.jabref.bst.BibtexCaseChanger.FORMAT_MODE;
+import net.sf.jabref.logic.bst.BibtexCaseChanger.FORMAT_MODE;
 
 import org.junit.Assert;
 import org.junit.Test;
