+++ src/test/resources/net/sf/jabref/logic/bst/abbrv.bst
